Re-organize commands into multiple files
